pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insightbot"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "Flask>=3.0",
    "pandas",
    "python-dotenv",
    "python-dateutil",
]
